feat: add schema and resolvers
