+++ apps/api/src/config/index.ts
@@ -1,0 +1,1 @@
+export const APP_SECRET = process.env.APP_SECRET || ''
