+++ apps/api/src/index.ts
@@ -1,9 +1,8 @@
+import { loadFilesSync } from '@graphql-tools/load-files'
+import { mergeResolvers, mergeTypeDefs } from '@graphql-tools/merge'
 import { ApolloServer } from 'apollo-server'
-import { readFileSync } from 'fs'
 import { join } from 'path'
 import { PrismaClient } from './generated/client'
-import { Resolvers } from './generated/types'
-import { Link, Mutation, Query, User } from './resolvers'
 import { getUserId } from './utils/getUserId'
 
 const context = ({ req }: { req: any }) => ({
@@ -11,18 +10,13 @@ const context = ({ req }: { req: any }) => ({
   prisma: new PrismaClient(),
   userId: req.headers.authorization ? getUserId(req) : null,
 })
-const resolvers: Resolvers = {
-  Link,
-  Mutation,
-  Query,
-  User,
-}
-const typeDefs = readFileSync(join(__dirname, 'schema.graphql'), 'utf8')
+const resolvers = loadFilesSync(join(__dirname, './resolvers'))
+const typeDefs = loadFilesSync(join(__dirname, './schema'))
 
 const server = new ApolloServer({
   context,
-  resolvers,
-  typeDefs,
+  resolvers: mergeResolvers(resolvers),
+  typeDefs: mergeTypeDefs(typeDefs),
 })
 
 server.listen().then(({ url }) => console.log(`🚀  Server ready at ${url}`))
