--- apps/api/src/resolvers/User.ts
@@ -1,8 +1,0 @@
-import { UserResolvers } from '../generated/types'
-
-const links: UserResolvers['links'] = (parent, _args, context) =>
-  context.prisma.user.findUnique({ where: { id: parent.id } }).links()
-
-export const User = {
-  links,
-}
