--- apps/api/src/resolvers/index.ts
@@ -1,4 +1,0 @@
-export { Link } from './Link'
-export { Mutation } from './Mutation'
-export { Query } from './Query'
-export { User } from './User'
