--- apps/api/src/schema.graphql
@@ -1,28 +1,0 @@
-type AuthPayload {
-  token: String
-  user: User
-}
-
-type Link {
-  description: String!
-  id: ID!
-  postedBy: User
-  url: String!
-}
-
-type Mutation {
-  login(email: String!, password: String!): AuthPayload
-  post(url: String!, description: String!): Link!
-  signup(email: String!, password: String!, name: String!): AuthPayload
-}
-
-type Query {
-  feed(filter: String, skip: Int, take: Int): [Link!]!
-}
-
-type User {
-  email: String!
-  id: ID!
-  links: [Link!]!
-  name: String!
-}
