+++ apps/api/src/schema/role.schema.graphql
@@ -1,0 +1,15 @@
+type Role {
+  id: ID!
+  name: String!
+}
+
+type Mutation {
+  createRole(name: String!): Role
+  updateRole(id: ID!, name: String!): Role
+  deleteRole(id: ID!): Role
+}
+
+type Query {
+  role(id: ID!): Role!
+  roles: [Role]!
+}
