+++ apps/api/src/schema/user.schema.graphql
@@ -1,0 +1,16 @@
+type AuthPayload {
+  token: String
+  user: User
+}
+
+type User {
+  id: ID!
+  name: String!
+  email: String!
+  role: Role!
+}
+
+type Mutation {
+  createUser(email: String!, name: String!, password: String!): AuthPayload
+  loginUser(email: String!, password: String!): AuthPayload
+}
